refactor me
